% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_baselines_one_location.R
\name{fit_baselines_one_location}
\alias{fit_baselines_one_location}
\title{Generate predictions for all baseline models, for a single location}
\usage{
fit_baselines_one_location(
  model_variations,
  target_ts,
  reference_date,
  temporal_resolution,
  horizons,
  quantile_levels,
  n_samples,
  round_predictions = FALSE,
  seed = NULL
)
}
\arguments{
\item{model_variations}{a \code{data.frame} where each row specifies a set of
hyperparameters to use for a single baseline model fit, with columns
\code{transformation}, \code{symmetrize}, and \code{window_size}. See details for more
information}

\item{target_ts}{a \code{data.frame} of target data in a time series format
(contains columns \code{time_index}, \code{location}, and \code{observation}) for a single
location}

\item{reference_date}{string of the reference date for the forecasts, i.e.
the date relative to which the targets are defined (usually Saturday for
weekly targets)}

\item{temporal_resolution}{'daily' or 'weekly'; specifies timescale of
\code{target_ts} and \code{horizons}}

\item{horizons}{numeric vector of prediction horizons relative to
the reference_date, e.g. 0:3 or 1:4}

\item{quantile_levels}{numeric vector of quantile levels to output; set to NULL
if quantile outputs not requested. Defaults to NULL.}

\item{n_samples}{integer of amount of samples to output (and predict);
set to NULL if sample outputs not requested (in this case 100000 samples
are generated from which to extract quantiles). Defaults to NULL.}

\item{round_predictions}{boolean specifying whether to round the output
predictions to the nearest whole number. Defaults to FALSE}

\item{seed}{integer specifying a seed to set for reproducible results.
Defaults to NULL, in which case no seed is set.}
}
\value{
data frame of a baseline forecast for one location, all models with
columns \code{transformation}, \code{symmetrize}, \code{window_size}, \code{horizon},
\code{output_type}, \code{output_type_id}, and \code{value}
}
\description{
Generate predictions for all baseline models, for a single location
}
\details{
The \code{model_variations} data frame has the following columns and
possible values for each:
\itemize{
\item transformation (character): "none" or "sqrt", determines distribution shape
\item symmetrize (boolean), determines if distribution is symmetric
\item window_size (integer), determines how many previous observations inform
the forecast
}

Additionally, this function will return slightly different output forecasts
depending on the relationship between the \code{reference_date}, requested \code{horizons},
and dates contained within \code{target_ts}. There are three possible cases:
\enumerate{
\item The requested forecasts begin exactly one time unit (given by the
temporal_resolution) after the last observed value. Here, no changes are
made to the returned forecasts.
\item The requested forecasts begin two or more time units after the last
observed value. Here, we make predictions starting from the next time unit
after the last observed value until the last requested date, returning
only forecasts for the requested dates.
\item The dates for the requested forecasts overlap partially or completely with
observed values contained within \code{target_ts}. Here, any forecasted values
for overlapping dates are replaced by the associated observed values.
(Note that we warn for this case.)
}
}
